End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -3,6 +3,9 @@ using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -16,9 +19,16 @@ namespace AuctionService.Controllers
         private readonly AuctionDbContext _context;
         private readonly IMapper _mapper;
 
-        public AuctionsController(AuctionDbContext context, IMapper mapper) { 
+        private IPublishEndpoint _publishEndpoint { get; }
+
+        public AuctionsController(
+            AuctionDbContext context,
+            IMapper mapper,
+            IPublishEndpoint publishEndpoint
+            ) { 
             _context = context;
             _mapper = mapper;
+            _publishEndpoint = publishEndpoint;
         }
 
         [HttpGet]
@@ -55,14 +65,19 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
+
+            var newAuction = _mapper.Map<AuctionDto>(auction);
+
+            await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
 
             var result  = await _context.SaveChangesAsync() > 0;
             if(!result)
@@ -70,13 +85,15 @@ namespace AuctionService.Controllers
                 return BadRequest("Could not save changes to DB");
             }
 
+
             return CreatedAtAction(
                 nameof(GetAuctionById),
                 new {auction.Id},
-                _mapper.Map<AuctionDto>(auction)
+                newAuction
                 );
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(UpdateAuctionDto updateAuctionDto,Guid id)
         {
@@ -89,11 +106,15 @@ namespace AuctionService.Controllers
                 return NotFound();
             }
 
+            if (auction.Seller != User.Identity.Name) return Forbid();
+
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
             auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
             auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
             auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
+
+            await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
 
             var result = await _context.SaveChangesAsync() > 0;
             if(result)
@@ -104,6 +125,7 @@ namespace AuctionService.Controllers
             return BadRequest("Failed updating");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuctionById(Guid id)
         {
@@ -115,7 +137,11 @@ namespace AuctionService.Controllers
                 return NotFound();
             }
 
+            if (auction.Seller != User.Identity.Name) return Forbid();
+
             _context.Auctions.Remove(auction);
+
+            await _publishEndpoint.Publish<AuctionDeleted>( new {Id = auction.Id.ToString()} );
             var result = await _context.SaveChangesAsync() > 0;
 
             if(!result)
